backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.3.3
numpy==2.3.5
scipy==1.16.3
joblib==1.5.2
xgboost==3.1.2
scikit-learn==1.7.2
requests==2.32.5
python-dotenv==1.1.0
